pyproject.toml
#####################################################################################
# A tool for the creation of JasperGold SVP principle tcl files.
# Copyright (C) 2024  RISCY-Lib Contributors
#
# This program is free software: you can redistribute it and/or modify
# it under the terms of the GNU General Public License as published by
# the Free Software Foundation, either version 3 of the License, or
# (at your option) any later version.
#
# This program is distributed in the hope that it will be useful,
# but WITHOUT ANY WARRANTY; without even the implied warranty of
# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
# GNU General Public License for more details.
#
# You should have received a copy of the GNU General Public License
# along with this program.  If not, see <https://www.gnu.org/licenses/>.
#####################################################################################

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mavsec"
dynamic = ["version"]
description = "A tool for the creation of JasperGold SVP principle tcl files."
readme = "README.md"
authors = [{name = "Benjamin Davis", email = "allrisc.dev@gmail.com"}]
license = {file = "LICENSE"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: GNU General Public License v3 or later (GPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Operating System :: OS Independent",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Utilities",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Electronic Design Automation (EDA)",
    "Environment :: X11 Applications :: Qt",
    "Typing :: Typed",
]
keywords = [
    "Verilog",
    "tool",
    "SystemVerilog",
    "JasperGold",
    "SVP",
    "Security",
    "Verification",
    "Formal",
    "Property",
    "Assertion"
]

dependencies = ["pyside6", "pyyaml", "tomli-w"]
requires-python = ">=3.8"

[project.urls]
Homepage = "https://github.com/RISCY-Lib/mavsec"
Documentation = "https://github.com/RISCY-Lib/mavsec/blob/main/README.md"

[project.scripts]
mavsec = "mavsec:gui"

[project.optional-dependencies]
dev = ["flake8", "mypy", "tox", "pytest", "setuptools", "sphinx", "sphinx-rtd-theme"]